@@ -6,7 +6,7 @@ name: front_end
 publish_to: none
 
 environment:
-  sdk: '>=2.17.0 <3.0.0'
+  sdk: ^3.0.0
 
 # Use 'any' constraints here; we get our versions from the DEPS file.
 dependencies:
