@@ -4,7 +4,7 @@ name: nnbd_migration
 publish_to: none
 
 environment:
-  sdk: '>=2.19.0-0 <3.0.0'
+  sdk: ^3.0.0
 
 # Use 'any' constraints here; we get our versions from the DEPS file.
 dependencies:
