@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-abstract class FactorTypeTestMixin<Type> {
+mixin FactorTypeTestMixin<Type> {
   Type futureNone(Type type);
   Type futureOrNone(Type type);
 
