@@ -5,7 +5,7 @@
 part of '../api.dart';
 
 /// The marker interface for all types of macros.
-abstract class Macro {}
+abstract mixin class Macro {}
 
 /// The interface for [Macro]s that can be applied to any top level function,
 /// instance method, or static method, and want to contribute new type
