@@ -80,18 +80,30 @@ class SourceFactoryBuilder extends SourceFunctionBuilderImpl {
       : super(metadata, modifiers, name, typeVariables, formals, libraryBuilder,
             charOffset, nativeMethodName) {
     _procedureInternal = new Procedure(
-        dummyName, ProcedureKind.Factory, new FunctionNode(null),
-        fileUri: libraryBuilder.fileUri, reference: procedureReference)
+        dummyName,
+        nameScheme.isInlineClassMember
+            ? ProcedureKind.Method
+            : ProcedureKind.Factory,
+        new FunctionNode(null),
+        fileUri: libraryBuilder.fileUri,
+        reference: procedureReference)
       ..fileStartOffset = startCharOffset
       ..fileOffset = charOffset
       ..fileEndOffset = charEndOffset
-      ..isNonNullableByDefault = libraryBuilder.isNonNullableByDefault;
+      ..isNonNullableByDefault = libraryBuilder.isNonNullableByDefault
+      ..isInlineClassMember = nameScheme.isInlineClassMember;
     nameScheme
         .getProcedureMemberName(ProcedureKind.Factory, name)
         .attachMember(_procedureInternal);
     _factoryTearOff = createFactoryTearOffProcedure(name, libraryBuilder,
-        libraryBuilder.fileUri, charOffset, tearOffReference);
-    // TODO(johnniwinther): Use [NameScheme] for constructor tear-off names.
+        libraryBuilder.fileUri, charOffset, tearOffReference,
+        forceCreateLowering: nameScheme.isInlineClassMember);
+    if (_factoryTearOff != null) {
+      _factoryTearOff!..isInlineClassMember = nameScheme.isInlineClassMember;
+      nameScheme
+          .getConstructorMemberName(name, isTearOff: true)
+          .attachMember(_factoryTearOff!);
+    }
     this.asyncModifier = asyncModifier;
   }
 
@@ -147,9 +159,17 @@ class SourceFactoryBuilder extends SourceFunctionBuilderImpl {
   @override
   void buildOutlineNodes(void Function(Member, BuiltMemberKind) f) {
     _build();
-    f(_procedureInternal, BuiltMemberKind.Method);
+    f(
+        _procedureInternal,
+        isInlineClassMember
+            ? BuiltMemberKind.InlineClassFactory
+            : BuiltMemberKind.Factory);
     if (_factoryTearOff != null) {
-      f(_factoryTearOff!, BuiltMemberKind.Method);
+      f(
+          _factoryTearOff!,
+          isInlineClassMember
+              ? BuiltMemberKind.InlineClassTearOff
+              : BuiltMemberKind.Method);
     }
   }
 
