@@ -1519,13 +1519,11 @@ void NativeReturnInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
   __ StoreToOffset(tmp, THR, compiler::target::Thread::top_resource_offset());
 
   // Reset the exit frame info to old_exit_frame_reg *before* entering the
-  // safepoint.
-  //
-  // If we were called by a trampoline, it will enter the safepoint on our
+  // safepoint. The trampoline that called us will enter the safepoint on our
   // behalf.
-  __ TransitionGeneratedToNative(
-      vm_tag_reg, old_exit_frame_reg, old_exit_through_ffi_reg,
-      /*enter_safepoint=*/!NativeCallbackTrampolines::Enabled());
+  __ TransitionGeneratedToNative(vm_tag_reg, old_exit_frame_reg,
+                                 old_exit_through_ffi_reg,
+                                 /*enter_safepoint=*/false);
 
   __ PopNativeCalleeSavedRegisters();
 
@@ -1573,24 +1571,6 @@ void NativeEntryInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
 
   __ PushNativeCalleeSavedRegisters();
 
-  // Load the thread object. If we were called by a JIT trampoline, the thread
-  // is already loaded.
-  const intptr_t callback_id = marshaller_.dart_signature().FfiCallbackId();
-  if (!NativeCallbackTrampolines::Enabled()) {
-    compiler->LoadBSSEntry(BSS::Relocation::DRT_GetThreadForNativeCallback, R1,
-                           R0);
-    // Create another frame to align the frame before continuing in "native"
-    // code.
-    __ EnterFrame(0);
-    __ ReserveAlignedFrameSpace(0);
-
-    __ LoadImmediate(R0, callback_id);
-    __ blr(R1);
-    __ mov(THR, R0);
-
-    __ LeaveFrame();
-  }
-
   // Now that we have THR, we can set CSP.
   __ SetupCSPFromThread(THR);
 
@@ -1627,14 +1607,15 @@ void NativeEntryInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
   // correct offset from FP.
   __ EmitEntryFrameVerification();
 
-  // Either DLRT_GetThreadForNativeCallback or the callback trampoline (caller)
-  // will leave the safepoint for us.
+  // The callback trampoline (caller) has already left the safepoint for us.
   __ TransitionNativeToGenerated(R0, /*exit_safepoint=*/false);
 
   // Now that the safepoint has ended, we can touch Dart objects without
   // handles.
 
   // Load the code object.
+  const Function& target_function = marshaller_.dart_signature();
+  const intptr_t callback_id = target_function.FfiCallbackId();
   __ LoadFromOffset(R0, THR, compiler::target::Thread::isolate_group_offset());
   __ LoadFromOffset(R0, R0,
                     compiler::target::IsolateGroup::object_store_offset());
