@@ -1276,7 +1276,8 @@ const char* ImageWriter::SnapshotTextObjectNamer::SnapshotNameFor(
 
 void AssemblyImageWriter::WriteBss(bool vm) {
   EnterSection(ProgramSection::Bss, vm, ImageWriter::kBssAlignment);
-  auto const entry_count = vm ? BSS::kVmEntryCount : BSS::kIsolateEntryCount;
+  auto const entry_count =
+      vm ? BSS::kVmEntryCount : BSS::kIsolateGroupEntryCount;
   for (intptr_t i = 0; i < entry_count; i++) {
     // All bytes in the .bss section must be zero.
     WriteTargetWord(0);
