@@ -47,7 +47,9 @@ VirtualMemory* VirtualMemory::ForImagePage(void* pointer, uword size) {
   return memory;
 }
 
-VirtualMemory* VirtualMemory::DuplicateRX() {
+bool VirtualMemory::DuplicateRX(VirtualMemory* target) {
+  ASSERT_LESS_OR_EQUAL(size(), target->size());
+
 #if defined(DART_HOST_OS_MACOS)
   // Mac is special cased because iOS doesn't allow allocating new executable
   // memory, so the default approach would fail. We are allowed to make new
@@ -56,33 +58,32 @@ VirtualMemory* VirtualMemory::DuplicateRX() {
   const mach_port_t task = mach_task_self();
   const vm_address_t source_address = reinterpret_cast<vm_address_t>(address());
   const vm_size_t mem_size = size();
-  vm_prot_t current_protection = VM_PROT_READ | VM_PROT_EXECUTE;
-  vm_prot_t max_protection = current_protection;
-  vm_address_t target_address = 0;
-  kern_return_t status =
-      vm_remap(task, &target_address, mem_size,
-               /*mask=*/0,
-               /*flags=anywhere*/ 1, task, source_address,
-               /*copy=*/true, &current_protection, &max_protection,
-               /*inheritance=*/VM_INHERIT_NONE);
+  const vm_prot_t read_execute = VM_PROT_READ | VM_PROT_EXECUTE;
+  vm_prot_t current_protection = read_execute;
+  vm_prot_t max_protection = read_execute;
+  vm_address_t target_address =
+      reinterpret_cast<vm_address_t>(target->address());
+  kern_return_t status = vm_remap(
+      task, &target_address, mem_size,
+      /*mask=*/0,
+      /*flags=*/VM_FLAGS_FIXED | VM_FLAGS_OVERWRITE, task, source_address,
+      /*copy=*/true, &current_protection, &max_protection,
+      /*inheritance=*/VM_INHERIT_NONE);
   if (status != KERN_SUCCESS) {
-    return nullptr;
+    return false;
   }
-  const MemoryRegion region(reinterpret_cast<void*>(target_address), mem_size);
-  return new VirtualMemory(region, region);
+  ASSERT(reinterpret_cast<void*>(target_address) == target->address());
+  ASSERT_EQUAL(current_protection & read_execute, read_execute);
+  ASSERT_EQUAL(max_protection & read_execute, read_execute);
+  return true;
 
 #else   // defined(DART_HOST_OS_MACOS)
-  // TODO(liama): Use dual mapping on platforms where it's supported.
-  VirtualMemory* const memory = VirtualMemory::Allocate(
-      size(), /*is_executable=*/true, /*is_compressed=*/false,
-      "VirtualMemory::DuplicateRX");
-  if (memory == nullptr) {
-    return nullptr;
-  }
-  // This is safe because the old memory can't overlap with the new memory.
-  memcpy(memory->address(), address(), size());  // NOLINT
-  memory->Protect(kReadExecute);
-  return memory;
+  // TODO(52497): Use dual mapping on platforms where it's supported.
+  // Check that target doesn't overlap with this.
+  ASSERT(target->start() >= end() || target->end() <= start());
+  memcpy(target->address(), address(), size());  // NOLINT
+  Protect(target->address(), size(), kReadExecute);
+  return true;
 #endif  // defined(DART_HOST_OS_MACOS)
 }
 
