@@ -63,10 +63,14 @@ class VirtualMemory {
                                         bool is_compressed,
                                         const char* name);
 
-  // Duplicates kReadExecute memory. This is designed to work on all platforms,
-  // including iOS, which doesn't allow creating new executable memory. The
-  // VirtualMemory being duplicated must have the kReadExecute protection level.
-  VirtualMemory* DuplicateRX();
+  // Duplicates `this` memory into the `target` memory. This is designed to work
+  // on all platforms, including iOS, which doesn't allow creating new
+  // executable memory.
+  //
+  // Assumes
+  //   * `this` has RX protection.
+  //   * `target` has RW protection, and is at least as large as `this`.
+  bool DuplicateRX(VirtualMemory* target);
 
   // Returns the cached page size. Use only if Init() has been called.
   static intptr_t PageSize() {
